autogen symlink script: ignore marker files

+++ scripts/build/autogen_ros_version_src.py
@@ -55,6 +55,13 @@ The transformation behavior is:
     out.  The Jinja2 rendering context contains the boolean variable
     ROS1 that can be used in Jinja2 conditionals to modify the content
     between ROS versions.
+  - Any folder that contains a "ROS1_IGNORE" or "ROS2_IGNORE" marker
+    file will not be propagated to the output folder if the name matches
+    the current ROS version. Note that this behavior is similar to the
+    ROS native "CATKIN_IGNORE" and "COLCON_IGNORE" marker files, but it
+    turns out that both ROS versions respect the ignore markers of the
+    other ROS version, which defeats our objective to have different
+    behavior between the ROS versions.
 
 This transformation has the following nice properties:
 - It is performing an out-of-source build such that the checkout folder
@@ -126,9 +133,7 @@ def get_current_state(autogen_dir):
     """
     result = {}
     for dir_path, dir_names, file_names in os.walk(autogen_dir, followlinks=False):
-        for d in dir_names:
-            autogen_d = os.path.join(dir_path, d)
-            result[autogen_d] = {"type": "d"}
+        result[dir_path] = {"type": "d"}
 
         for f in file_names:
             autogen_f = os.path.join(dir_path, f)
@@ -164,15 +169,19 @@ def get_desired_state(checkout_dir, autogen_dir, use_ros1):
     mapping paths under autogen_dir to their desired type and content.
     """
     result = {}
+    ignore_marker = "ROS1_IGNORE" if use_ros1 else "ROS2_IGNORE"
     for dir_path, dir_names, file_names in os.walk(checkout_dir, followlinks=True):
-        if dir_path == checkout_dir and ".git" in dir_names:
-            dir_names.remove(".git")  # skip .git subfolder
+        if ignore_marker in file_names:
+            # if folder contains ignore_marker file, ignore it and all of its
+            # children
+            dir_names[:] = []
+            continue
 
-        for d in dir_names:
-            autogen_d = os.path.join(
-                autogen_path(dir_path, checkout_dir, autogen_dir), d
-            )
-            result[autogen_d] = {"type": "d"}
+        if ".git" in dir_names:
+            dir_names.remove(".git")  # skip any .git subfolder
+
+        autogen_dir_path = autogen_path(dir_path, checkout_dir, autogen_dir)
+        result[autogen_dir_path] = {"type": "d"}
 
         for f in file_names:
             checkout_f = os.path.join(dir_path, f)
@@ -260,10 +269,11 @@ def apply_update_map(update_map, dry_run):
         desired = val["desired"]
 
         if current is not None:
-            if current["type"] == "d":
-                shutil.rmtree(p)
-            else:
-                os.unlink(p)
+            if os.path.exists(p):
+                if current["type"] == "d":
+                    shutil.rmtree(p)
+                else:
+                    os.unlink(p)
 
         if desired is not None:
             if not os.path.exists(os.path.dirname(p)):
